Drop org maintainers team (#227)
The intent here isn't clear. I think it's supposed to be "full
access", but that is a separate configuration (github org owners) and
many of these same folks are also github org owners.  Permissiosn
given to this team is also not consistent across repos.

Just drop it and focus on setting github org owners as the ones that
need full access for admin purposes. I also don't think it's important
to list that group here.

If we merge this, I will remove the corresponding github team, as
well.

Signed-off-by: Russell Bryant <rbryant@redhat.com>

Signed-off-by: Russell Bryant <rbryant@redhat.com>