Reflect the bootstrapping oversight committee
The governance doc discussed an oversight committee, but one had not
yet been created. This change reflects an initial bootstrapping
oversight committee appointed at the launch of the project that will
serve in this function as the project gets started. One of the roles
of this group will be to finalize the governance processes for this
committee for the future.

While in the doc anyway, I did some other minor cleanups - minor
grammatical issues and references to things that don't exist.

Signed-off-by: Russell Bryant <rbryant@redhat.com>