Clarify that O.C. members have authority over team membership
We have had a couple recent cases where a member of the oversight
committee saw a need for adjustments to team membership and made those
changes without going through the formalities discussed in our
CONTRIBUTOR_ROLES document.

We believe that is acceptable at this early stage of the
project. Update the document to make it clear that this is an expected
and approved part of our governance at this time.

Signed-off-by: Russell Bryant <rbryant@redhat.com>