Require Maintainers to be Triagers
The project is looking for ways to improve community relationships, and
one of the suggestions is to require maintainers to share triager
responsibilities to provide timely and useful feedback to contributors
on their issues and pull requests.

The proposal is to assume privileges and responsibilities and membership
is monotonically expanding so that:

- all triagers are members (already the case);
- all maintainers are triagers (this proposal).

Or if these were python sets, then... ;)

```
Contributor <= Member <= Triager <= Maintainer
```

Context: we are considering automatic assignment of a number of Triagers
to every new pull request, and it's important that this burden is shared
by all seasoned members, incl. maintainers.

Note: once this is approved, someone should walk through the list of
current Maintainers and add them to Triagers groups for appropriate
repos. Alternatively, Maintainers groups can be converted into nested
groups, see:

https://docs.github.com/en/organizations/organizing-members-into-teams/about-teams#nested-teams

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>